+++ subsys/bluetooth/controller/ll_sw/ull_df.c
@@ -25,7 +25,7 @@
 #include "ull_adv_internal.h"
 
 #include "ull_df.h"
-#include "lll_df.h"
+#include "lll_df_internal.h"
 
 #define BT_DBG_ENABLED IS_ENABLED(CONFIG_BT_DEBUG_HCI_DRIVER)
 #define LOG_MODULE_NAME bt_ctlr_ull_df
